#68 Copy plugins to save during the volume migration
The plugin volume mounts into a directory where files might already exist. That
would be the case if the dogu was installed in an earlier version. The new
volume can't just "over-mount" the files: Either the files would disappear or
Docker would throw an error.

The solution to this problem is a two-phase move:
1. Move away all plugins from the original filesystem during the pre-upgrade
2. Mount plugins volume
3. Move back all plugins to the plugins volume during the post-upgrade