#82 Fix incompatible semantic version comparison
Up to year 2020 `sort -v` was not available in Alpine bash, and so, upgrading
from an older Redmine version (f. i. 4.0.5-1) failed horrifically, leaving the
Redmine dogu in an indeterminate state.

This commit removes `sort -v` in favor of a bash regex implementation. Some
SemVer functionality is lost with this commit (f. i. 1.0.0 < 1.0.0-1 etc)
but the used version scheme is under tight development control so
"uncanonical" versions should not be used.