#130 Fix markdown formatter migration in post-upgrade
The post-upgrade bash function contained a dangerous bug which migrated
every formatter entry to "common_mark" (not only "markdown"). This
commit extracts the whole logic adds, fixes the logic to pin down only
the markdown formatter, and adds some log output to add transparency
during the post-upgrade.

Additional BATS tests help in understanding when and how the migration
is executed.

Co-authored-by: Georg Goering <georg.goering-extern@cloudogu.com>