#130 disable markdown formatter migration for warning
Previous markdown tests missed the point that `_underlined_` markdown no
longer worked in the new formatter. This creates a too strong change
for an inconsiderate migration. Because the formatter `markdown` is
still available and not yet removed, this commit replaces the migration
routine with strong warnings in all places to take care of the upcoming
formatter change.

The actual migration must be used when Redmine actually removes the
formatter from its formatter list.