#130 Fix ANSI color in upgrade notification, add missing method
The function versionXLessOrEqualThanY() was only available in
pre-upgrade.sh. This commit copies the same function to the
upgrade-notification.sh in order to be independently available.

Regarding the printf color syntax see also this SO reply which explains
that a simple variable assignment of ANSI escapes does not work but must
be interpreted instead (by printf or tput):

https://unix.stackexchange.com/questions/266921/is-it-possible-to-use-ansi-color-escape-codes-in-bash-here-documents