#143 Update default_config key behaviour to comply with CES MN
With MN CES there will no longer exist a way to update or delete config
values for config keys that come from outside. These keys are mounted as
readonly files.

This commit adds logic to emulate the previous behaviour of skipping the
import of default data IF IT HAS CHANGED. To continue the support of
auditable settings, a new key will accommodate the last import timestamp
to look up the last imported data. If both match no further import is
deemed necessary. In all other cases where a value is non-empty, the
data are tried to be parsed and imported as in earlier releases.